move: move Product.java from /gift to /gift/model
