+++ src/main/java/gift/model/ProductForm.java
@@ -1,0 +1,2 @@
+package gift.model;public class ProductForm {
+}
